Stop relying on transitive includes and an implicit narrowing
Two portability defects found by auditing the Windows-only sources rather
than waiting for a compiler to reach them. The previous CI build stopped
at 95 errors before it got this far, so neither had been exercised.

remote_heap_probe.cpp calls std::toupper without <cctype> and
std::make_unique without <memory>. It builds on Linux only because other
headers in the translation unit happen to pull both in, which is not
something to depend on.

meminfo_hook.cpp passed sizeof(buf) -- a 64-bit size_t -- as
GetEnvironmentVariableA's 32-bit DWORD parameter. MSVC reports the
narrowing as C4267 and this project compiles with /WX, so it would have
failed the build. The cast is now explicit.

61/61 pass on Linux; the hook still syntax-checks clean against the stub
Win32 headers.

+++ hook/meminfo_hook.cpp
@@ -95,7 +95,9 @@ void log_line(const char* fmt, ...) {
 
 std::string env_string(const char* name, const std::string& fallback) {
     char buf[1024];
-    const DWORD n = GetEnvironmentVariableA(name, buf, sizeof(buf));
+    // sizeof is size_t; the parameter is a 32-bit DWORD. The cast is explicit
+    // because MSVC reports the narrowing as C4267 and /WX makes it fatal.
+    const DWORD n = GetEnvironmentVariableA(name, buf, static_cast<DWORD>(sizeof(buf)));
     if (n == 0 || n >= sizeof(buf)) return fallback;
     return std::string(buf, n);
 }
